+++ matematikai_fuggvenyek.h
@@ -11,6 +11,6 @@ bool vissza_lep(sf::Event& e);
 
 bool palyan_belul(int x, int y);
 
-size_t eger_lekepezes(size_t mit, size_t eltolas, size_t mennyivel);
+unsigned long long eger_lekepezes(int mit, int eltolas, int mennyivel);
 
 bool valid_idx(int idx);
